fix(Input): 정수 범위 벗어난 입력 예외 처리, 패키지명 변경
- io -> view로 패키지명 변경

inputPruchaseAmount
- try catch를 통해 NumberFormatException 처리

InstructionMessages
- enum 오타 수정
