+++ src/main/java/lotto/view/InstructionMessages.java
@@ -1,8 +1,8 @@
-package lotto.io;
+package lotto.view;
 
 public enum InstructionMessages {
     INPUT_PURCHASE_AMOUNT("구입금액을 입력해 주세요."),
-    INPUT_WINNINT_NUMBERS("당첨 번호를 입력해 주세요."),
+    INPUT_WINNING_NUMBERS("당첨 번호를 입력해 주세요."),
     INPUT_BONUS_NUMBER("보너스 번호를 입력해 주세요."),
     NUMBER_OF_PURCHASE("개를 구매했습니다."),
     WINNING_STATISTICS("당첨 통계")
